+++ pooi_project/src/main/java/entity/CareerType.java
