--- pooi_project/src/service/foo
@@ -1,1 +1,0 @@
-Falando besteiras.
